+++ src/video/win32/pal_video_win32.h
@@ -81,7 +81,6 @@ typedef struct {
 
 typedef struct {
     PalBool initialized;
-    int32_t pixelFormat;
     int32_t maxWindowData;
     PalVideoFeatures features;
     const PalAllocator* allocator;