+++ tools/abi_dump/video_abi_dump.c
@@ -324,7 +324,7 @@ static void windowInfoDump()
 
 static void windowDump()
 {
-    uint32_t xSize = 64;
+    uint32_t xSize = 72;
     uint32_t xAlign = 8;
     uint32_t xOffset1 = 0;
     uint32_t xOffset2 = 8;
@@ -337,6 +337,8 @@ static void windowDump()
     uint32_t xOffset9 = 52;
     uint32_t xOffset10 = 56;
     uint32_t xOffset11 = 60;
+    uint32_t xOffset12 = 64;
+    uint32_t xOffset13 = 68;
     uint32_t xPadding = 0;
 
     uint32_t ySize = sizeof(PalWindowCreateInfo);
@@ -346,12 +348,14 @@ static void windowDump()
     uint32_t yOffset3 = offsetof(PalWindowCreateInfo, monitor);
     uint32_t yOffset4 = offsetof(PalWindowCreateInfo, appName);
     uint32_t yOffset5 = offsetof(PalWindowCreateInfo, instanceName);
-    uint32_t yOffset6 = offsetof(PalWindowCreateInfo, width);
-    uint32_t yOffset7 = offsetof(PalWindowCreateInfo, height);
-    uint32_t yOffset8 = offsetof(PalWindowCreateInfo, show);
-    uint32_t yOffset9 = offsetof(PalWindowCreateInfo, maximized);
-    uint32_t yOffset10 = offsetof(PalWindowCreateInfo, minimized);
-    uint32_t yOffset11 = offsetof(PalWindowCreateInfo, center);
+    uint32_t yOffset6 = offsetof(PalWindowCreateInfo, fbConfigBackend);
+    uint32_t yOffset7 = offsetof(PalWindowCreateInfo, fbConfigId);
+    uint32_t yOffset8 = offsetof(PalWindowCreateInfo, width);
+    uint32_t yOffset9 = offsetof(PalWindowCreateInfo, height);
+    uint32_t yOffset10 = offsetof(PalWindowCreateInfo, show);
+    uint32_t yOffset11 = offsetof(PalWindowCreateInfo, maximized);
+    uint32_t yOffset12 = offsetof(PalWindowCreateInfo, minimized);
+    uint32_t yOffset13 = offsetof(PalWindowCreateInfo, center);
     uint32_t yPadding = (yAlign - (ySize % yAlign)) % yAlign;
 
     const char* result = s_FailedString;
@@ -369,6 +373,8 @@ static void windowDump()
         xOffset9 == yOffset9      &&
         xOffset10 == yOffset10    &&
         xOffset11 == yOffset11    &&
+        xOffset12 == yOffset12    &&
+        xOffset13 == yOffset13    &&
         xPadding == yPadding) {
         result = s_PassedString;
     }
@@ -376,23 +382,25 @@ static void windowDump()
 
     palLog(nullptr, "PalWindowCreateInfo");
     palLog(nullptr, "===========================================");
-    palLog(nullptr, "Field          Expected     Actual");
+    palLog(nullptr, "Field            Expected     Actual");
     palLog(nullptr, "===========================================");
 
-    palLog(nullptr, "size           %u           %u", xSize, ySize);
-    palLog(nullptr, "align          %u            %u", xAlign, yAlign);
-    palLog(nullptr, "padding        %u            %u", xPadding, yPadding);
-    palLog(nullptr, "style @        %u            %u", xOffset1, yOffset1);
-    palLog(nullptr, "title @        %u            %u", xOffset2, yOffset2);
-    palLog(nullptr, "monitor @      %u           %u", xOffset3, yOffset3);
-    palLog(nullptr, "appName @      %u           %u", xOffset4, yOffset4);
-    palLog(nullptr, "instanceName @ %u           %u", xOffset5, yOffset5);
-    palLog(nullptr, "width @        %u           %u", xOffset6, yOffset6);
-    palLog(nullptr, "height @       %u           %u", xOffset7, yOffset7);
-    palLog(nullptr, "show @         %u           %u", xOffset8, yOffset8);
-    palLog(nullptr, "maximized @    %u           %u", xOffset9, yOffset9);
-    palLog(nullptr, "minimized @    %u           %u", xOffset10, yOffset10);
-    palLog(nullptr, "center @       %u           %u", xOffset11, yOffset11);
+    palLog(nullptr, "size               %u           %u", xSize, ySize);
+    palLog(nullptr, "align              %u            %u", xAlign, yAlign);
+    palLog(nullptr, "padding            %u            %u", xPadding, yPadding);
+    palLog(nullptr, "style @            %u            %u", xOffset1, yOffset1);
+    palLog(nullptr, "title @            %u            %u", xOffset2, yOffset2);
+    palLog(nullptr, "monitor @          %u           %u", xOffset3, yOffset3);
+    palLog(nullptr, "appName @          %u           %u", xOffset4, yOffset4);
+    palLog(nullptr, "instanceName @     %u           %u", xOffset5, yOffset5);
+    palLog(nullptr, "fbConfigBackend @  %u           %u", xOffset6, yOffset6);
+    palLog(nullptr, "fbConfigId @       %u           %u", xOffset7, yOffset7);
+    palLog(nullptr, "width @            %u           %u", xOffset8, yOffset8);
+    palLog(nullptr, "height @           %u           %u", xOffset9, yOffset9);
+    palLog(nullptr, "show @             %u           %u", xOffset10, yOffset10);
+    palLog(nullptr, "maximized @        %u           %u", xOffset11, yOffset11);
+    palLog(nullptr, "minimized @        %u           %u", xOffset12, yOffset12);
+    palLog(nullptr, "center @           %u           %u", xOffset13, yOffset13);
     palLog(nullptr, "===========================================");
 
     palLog(nullptr, "Status: %s", result);